Refactor: split utils.ts into focused modules
Complete the decomposition started in the guards/logger split:
- files.ts   — readFile/writeFile
- content.ts — code-fence masking, title extraction, partial imports,
               content cleaning, markdown template assembly
- images.ts  — image asset map + relative-image-URL rewriting

utils.ts drops from 879 to 385 lines (path/fs-listing/transform helpers +
joinSiteUrl) and re-exports every module so existing ./utils imports and
tests keep working. Modules are acyclically layered (guards/logger/files
are leaves; content depends on them; images depends on content). Removed
now-unused matter/YAML/Dirent imports from utils.

Behavior-preserving; build, full suite, and a real-plugin smoke test pass.
